ConsoleModeCommand and tests moved to seperate assemblies
