+++ ManyConsoleModeCommand/Internal/CommandLineParser.cs
