+++ ManyConsoleModeCommand/Properties/AssemblyInfo.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+#if !SILVERLIGHT
+[assembly: SuppressIldasmAttribute()]
+#endif
+[assembly: CLSCompliantAttribute(false )]
+[assembly: ComVisibleAttribute(false)]
+[assembly: AssemblyTitleAttribute("ManyConsoleModeCommand 1.0.0.3")]
+[assembly: AssemblyDescriptionAttribute("A library for writing console applications.  Extends Mono.Options to support separate commands from one console application.")]
+[assembly: AssemblyCompanyAttribute("n/a")]
+[assembly: AssemblyProductAttribute("ManyConsole 1.0.0.3")]
+[assembly: AssemblyCopyrightAttribute("Copyright Â© Frank Schwieterman 2011")]
+[assembly: AssemblyVersionAttribute("1.0.0.3")]
+//[assembly: AssemblyInformationalVersionAttribute("1.0.0.3 / 4efcc5b")]
+[assembly: AssemblyFileVersionAttribute("1.0.0.3")]
+[assembly: AssemblyDelaySignAttribute(false)]
+
